requirements-dev.txt
pytest
pytest-django
model_bakery